Use updated url to access s3 resources
Since aws-sdk@2.150.0 the urls have used a dot-separation between `s3` and `eu-west-1` (or other configured region). Since this is signed into the signature of the url it needs to match that used by the SDK internally.

Fix the version of `aws-sdk` to protect against future changes in AWS url schema.